factor out createInitialProfile__ utils
